Backed out 4 changesets (bug 1766377) for causing mochitest failures on HyperTextAccessible.cpp CLOSED TREE
Backed out changeset 6335511477a4 (bug 1766377)
Backed out changeset e8108d6fab9f (bug 1766377)
Backed out changeset 80d16d55c490 (bug 1766377)
Backed out changeset 1a6583d6e52e (bug 1766377)

UltraBlame original commit: ca68786294025e0cc0fc5123530233fce44b46ba

+++ browser/components/shell/nsWindowsShellService.cpp
@@ -1272,7 +1272,7 @@ static nsresult PinCurrentAppToTaskbarImpl(bool aCheckOnly,
   int shortcutCSIDLs[] = {CSIDL_COMMON_PROGRAMS, CSIDL_PROGRAMS,
                           CSIDL_COMMON_DESKTOPDIRECTORY,
                           CSIDL_DESKTOPDIRECTORY};
-  for (size_t i = 0; i < ArrayLength(shortcutCSIDLs); ++i) {
+  for (int i = 0; i < ArrayLength(shortcutCSIDLs); ++i) {
     
     
     
@@ -1539,7 +1539,7 @@ nsWindowsShellService::ClassifyShortcut(const nsAString& aPath,
                  {FOLDERID_UserPinned, u"\\TaskBar\\", u"Taskbar"},
                  {FOLDERID_UserPinned, u"\\StartMenu\\", u"StartMenu"}};
 
-  for (size_t i = 0; i < ArrayLength(folders); ++i) {
+  for (int i = 0; i < ArrayLength(folders); ++i) {
     nsAutoString knownPath;
